This for discussion only and not a final solution.
Resolves an issue with zeros getting stored against the m_AssetId in Unity 2018.3 in the new Prefab mode editor. I have also kept 2017 backward compatibility.

Its currently using the Experimental function to get the current prefab stage in 2018.3 - this should be updated when it moves out of experimental.

I have had to update UnityEditor.dll to 2018.3.0b8 so the new bindings are available. It still works with 2017 and doesnt cause errors as I make sure we dont call new methods if on a older version. It was throwing errors however in 2017 if the new methods were in a called methods even if those lines were not executed so I had to move those methods into a separate method only called in 2018.3. The method names are temporary and up for suggestion