Fix an error when trying to resolve an IP address
Alkanov had an error when trying to connect to an IP address. When passing an IP address into Dns.GetHostEntryAsync, it first looks up a reverse DNS for that IP. If that successful it then does a further resolve and returns all ip addresses for that reverse DNS result. In his case the reverse dns name didnt resolve back to his IP address.

The fix first tries to check if the host is already a valid IPAddress and just return it which will be quicker and prevent this error. I also allowed IPV6 as I presume thats also fine.