Add DNS Resolver to StartClient
Note the two deprecated methods of Dns on lines 333 and 343.  They've been marked deprecated for years but MS will not remove them because they know GetHostEntry is flaky in some cases, and in those cases the deprecated methods succeed where GetHostEntry fails.  Certainly they can be taken out and let GetHostEntry succeed or fail on its own.  That's open for discussion, obviously.