Add Mirror.Components AsmDef
This is part of the solution to Visual Studio not being able to hit breakpoints when debugging.  The other part the users have to do, and that's cover all their code with one or more AsmDef's that references both Mirror and Mirror.Components.

They can probably just create one in the project root (directly under Assets) that will suffice.