Update SyncVarHook.md
It seems the line "cachedMaterial.color = color;" is not working in the Universal Render Pipeline.
I had to add this line of code otherwise it would not work. 
In my version I have deleted the old line of code.
Here I have just commented it out as people with old/new versions can switch between or atleast try them.