changing BasicAuthenticator to use a serverCode
Using 2 fields doesn't make sense for this example as both have to be the same values as the servers value.

using a password is not a good idea as it implies to the user that data is secure when it is sent.
Instead use a simple code and say that the intended use is a private code that can be shared with friends.

re-ordering the methods so that all the server and client stuff are grouped.

Adding a note about mirror connections not being encrypted so to not send high security data in a normal message.