Calling OnStopServer on disconnect
This seems to be calling the OnStopServer but with a little twist add to it. It calls it twice which is unacceptable. I can add a private bool to it to check if it is called before but I believe I can find a better solution than a cheessy hack.

I am sending this commit just incase I forgot what I have done cuz I'm a little fish brained. I forgot shit real quick.