It is working as intented right now
I tried at least 4 different solutions by now and this seems to be the best way to handle it.

Underlying issue was:

I have been looking at this issue. OnStopServer is being called properly on server if it is server only.
On host server it is calling OnStopServer without an issue except only the local player itself.
So the underlying issue seems to be at local client on host server.

So I called OnStopServer during RemoveLocalConnection.