feat: Headless Benchmark (#394)
The benchmark can be run in editor using the "editorArgs" or it can be run in a standalone using normal command line args. Example command line arg usage:

```
.\MirrorNG.exe -transport tcp -server -monster 100 -client 10
```

In its current design at the time of this comment the specified players and monsters will constantly translate and rotate to provide network load. There is also a SyncVar float that will constantly update its value with a random number.

In future changes I would like to make the frequency of the objects moving adjustable and the number of syncvar/rpcs also adjustable to better simulate real gameplay.