fix: Remove MMO wording in readme + slight reword
Remove MMO wording, make it more generic so the average developer looking for Mirror can go "Nice, I can use this for my game which is a turn based RPG or card game or racer". Also remove advertising to uMMORPG and Cubica, it's pretty obvious that uMMORPG is by the same guy as one of the project leads. Dunno what's up with Cubica - probably that's better for the Mirror showcase.