refactor: NetworkIdentity syncvar wrapper (#512)
NetworkIdentities are a special case to use as syncvars.  There are cases where a network identity syncvar has been assigned an object that has not been spawned in the client, either because of unpredictable spawn order or because it is too far and area of interest culled it.

Before this PR,  the weaver did some crazy gymnastics to support those use cases.  This PR puts the logic in a struct,  the weaver will replace the syncvar field with this struct, and invoke the corresponding setter and getter instead.
This way the weaver can use the same code path for NetworkIdentities and other syncvars.

In the future we can do the same for NetworkBehaviors, Transforms and GameObjects.