adding snapshot if empty, and removing control of buffer from inspector
This will better handle stopping and starting movment.
If no snapshots add one at syncInterval before new snapshot with current position.
It is assumed that if there is no snapshot then the object wasn't moving.
We want to add a snapshot with current position to use an accurate time, if we dont then the lerp will jump when starting to move as time will be closer to new position.