feat: notify algorithm in connection (#549)
From the receiver it is the same as any other message:
```cs
   connection.RegisterHandler<SomeMessage>( (message) => {
       // I received a message 
   });
```

From the sender, it looks like this:
```cs
  connection.NotifyDelivered += (connection, token) => {
      Debug.Log($"the message with token {token} has been delivered");
  };
  connection.NotifyLost += (connection, token) => {
      Debug.Log($"the message with token {token} has been lost");
  };

  connection.SendNotify(new SomeMessage(...),  "my token");
```

If you send a message,  you will eventually get one of those callbacks.  Messages are always delivered in order.  There is no guarantee that messages are delivered (but you can just resend in the lost callback if you want).  Same latency as raw unreliable.