DontDestroyOnLoad gameobjects now won't remove observers during scene change
RemoveObservers in NetworkConnection now has a parameter "removeDontDestroyOnLoadObjects".
By default it is set to true. If set to false, gameObjects that are part of DontDestroyOnLoad
won't be removed from the observing set.
Because of that, updates of NetworkIdentities which have DontDestroyOnLoad activated,
will now not be lost for unready clients during scene changes.