Update NetworkServer.cs
NetworkIdentity.OnStartClient is not called when spawned when player object is not null. This is why network behaviours do not work properly because isClient is not set.

The proposed code fixes the problem by sending object spawn finished message to client if player object does not exist so isSpawnFinished variable on the client can be set to true.