fix: #3550 NetworkBehaviour on ... requires a NetworkIdentity.
fix for issue reported in: Error "Mirror.NetworkTransform on Turret requires a NetworkIdentity. Please add a NetworkIdentity component to Turret or it's parents." in clean project #3550

somehow UNITY_2021_OR_NEWER is evaluated to false where it should be true.
This fixes error message in UnityEditor
EditorTests are fine, some Runtime tests on my manchine are failing, but only same as before.