Add boolean to optionally keep the Network Manager in DDOL
Especially for EOSTransport, without this, it breaks the login sequence. When you are in a lobby, you're logged in, but when you disconnect, the current Network Manager is moved out of DDOL. And because you are loading into the main menu, it creates a brand new one and deletes the current one, which breaks all sorts of things in the EOS SDK. It would be great if you could add this bool, Thanks!